requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Django==4.1.4
idna==3.4
joblib==1.2.0
mysql==0.0.3
mysqlclient==2.1.1
nltk==3.8
Pillow==9.3.0
python-dateutil==2.8.2
qrcode==7.3.1
regex==2022.10.31
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
textblob==0.17.1
tqdm==4.64.1
tzdata==2022.7
urllib3==1.26.13
vaderSentiment==3.3.2
